ISSUE-2351 Run all browser tests with and without polyfill
remove unnecessary polyfill loading

scripts that need polyfill import themselves + unnecessary mocking removal

separate tests that require polyfill from tests that do not

remove unneeded test

web test runner create groups for tests to run with and without polyfills

refactor: rename
