+++ packages/ui/components/core/test/SlotMixin.test.js
@@ -1,30 +1,10 @@
-import sinon from 'sinon';
 import { defineCE, expect, fixture, fixtureSync, unsafeStatic, html } from '@open-wc/testing';
-import { ScopedElementsMixin } from '@open-wc/scoped-elements/lit-element.js';
 import { SlotMixin } from '@lion/ui/core.js';
 import { LitElement } from 'lit';
 
 /**
  * @typedef {import('../types/SlotMixinTypes.js').SlotHost} SlotHost
  */
-
-// @ts-ignore
-const createElementNative = ShadowRoot.prototype.createElement;
-function mockScopedRegistry() {
-  const outputObj = { createElementCallCount: 0 };
-  // @ts-expect-error wait for browser support
-  ShadowRoot.prototype.createElement = (tagName, options) => {
-    outputObj.createElementCallCount += 1;
-    // Return an element that lit can use as render target
-    return createElementNative(tagName, options);
-  };
-  return outputObj;
-}
-
-function unMockScopedRegistry() {
-  // @ts-expect-error wait for browser support
-  ShadowRoot.prototype.createElement = createElementNative;
-}
 
 describe('SlotMixin', () => {
   it('inserts provided element into light dom and sets slot', async () => {
@@ -564,88 +544,6 @@ describe('SlotMixin', () => {
 
       expect(varThatProvesAfterRenderIsCalled).to.equal('called');
       expect(el._templateNode.textContent).to.equal('1');
-    });
-  });
-
-  describe('Scoped Registries', () => {
-    it('supports scoped elements when polyfill loaded', async () => {
-      const outputObj = mockScopedRegistry();
-
-      class ScopedEl extends LitElement {}
-
-      const tagName = defineCE(
-        // @ts-ignore
-        class extends ScopedElementsMixin(SlotMixin(LitElement)) {
-          static get scopedElements() {
-            return {
-              // @ts-expect-error
-              ...super.scopedElements,
-              'scoped-elm': ScopedEl,
-            };
-          }
-
-          get slots() {
-            return {
-              ...super.slots,
-              template: () => html`<scoped-elm></scoped-elm>`,
-            };
-          }
-
-          render() {
-            return html`<slot name="template"></slot>`;
-          }
-        },
-      );
-
-      const tag = unsafeStatic(tagName);
-      await fixture(html`<${tag}></${tag}>`);
-
-      expect(outputObj.createElementCallCount).to.equal(1);
-
-      unMockScopedRegistry();
-    });
-
-    it('does not scope elements when polyfill not loaded', async () => {
-      // @ts-expect-error
-      ShadowRoot.prototype.createElement = null;
-      class ScopedEl extends LitElement {}
-
-      const tagName = defineCE(
-        class extends ScopedElementsMixin(SlotMixin(LitElement)) {
-          static get scopedElements() {
-            return {
-              // @ts-expect-error
-              ...super.scopedElements,
-              'scoped-el': ScopedEl,
-            };
-          }
-
-          get slots() {
-            return {
-              ...super.slots,
-              template: () => html`<scoped-el></scoped-el>`,
-            };
-          }
-
-          render() {
-            return html`<slot name="template"></slot>`;
-          }
-        },
-      );
-
-      const renderTarget = document.createElement('div');
-      const el = document.createElement(tagName);
-
-      // We don't use fixture, so we limit the amount of calls to document.createElement
-      const docSpy = sinon.spy(document, 'createElement');
-      document.body.appendChild(renderTarget);
-      renderTarget.appendChild(el);
-
-      expect(docSpy.callCount).to.equal(2);
-
-      document.body.removeChild(renderTarget);
-      docSpy.restore();
-      unMockScopedRegistry();
     });
   });
 });
